refactor(provider): split monolithic on_provider.go by concern
The provider-selector overlay lived in a single ~1.5k-line file. Split it
into cohesive, convention-named siblings (on_provider_<aspect>.go), matching
the existing on_plugin/on_mcp/on_identity layout:

  on_provider.go              overlay state + message routing (entry)
  on_provider_types.go        data model, messages, status-display helpers
  on_provider_nav.go          cursor/tab navigation, model search, key routing
  on_provider_select.go       model & provider selection
  on_provider_credentials.go  API-key entry, credential edit/remove
  on_provider_data.go         data loading, visible-list building, lifecycle
  on_provider_connect.go      connect/refresh flow and the in-flight spinner

Pure code move: the set of declarations and all behavior are unchanged.
